refactor: use function to generate sample story data
